refactor(task): relocate EnsureModelDownloadedTask to src/task/model and name the file after the task
The download task was the one task file not named after its task and not living
under src/task/. Rename src/config/ensureModelDownloaded.ts to
src/task/model/EnsureModelDownloadedTask.ts (and its test alongside), matching the
convention every other task file already follows, and move it out of the DI-config
folder into a new model/ task domain. Update the internal secModelRecord import and
all nine importers to the new path; no behavior change.
